Remove JSX from index.js in addition to pragma
This is an opinionated change, but having the main `index.js` require a jsx compilation step is just unnecessary for such a small project. Either there should be a `dist` folder with a fully compiled version or, as I am proposing, just remove the JSX and use the compiled version for `index.js`.

Fixes #6.